Resolved CID 351727 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->mVendors.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -1210,7 +1210,7 @@ void ObjectMgr::LoadVendors()
 
         delete result;
     }
-    LogDetail("ObjectMgr : %u vendors loaded.", mVendors.size());
+    LogDetail("ObjectMgr : %u vendors loaded.", static_cast<uint32_t>(mVendors.size()));
 }
 
 void ObjectMgr::ReloadVendors()
